♻️ refactor: 불필요한 client 호출 삭제
